Example enhanced Purple Air Synoptic dataset

The example_pas dataset provides a quickly loadable version of a pas object for practicing and code examples. This dataset contains data for sensors in Washington and Oregon and was generated on 2022-05-03 by running:

library(AirSensor2)

initializeMazamaSpatialUtils()

example_pas <-
  pas_createNew(
    apiReadKey = API_READ_KEY,
    countryCodes = "US",
    stateCodes = c("WA", "OR"),
    counties = NULL,
    lookbackDays = 1,
    outsideOnly = TRUE
  )

save(example_pas, file = "data/example_pas.rda")

example_pas

Format

A tibble with 1481 rows and 57 columns of data.

Source

https://www2.purpleair.com

See also