Example raw Purple Air Synoptic dataset

The example_pas_raw dataset provides a quickly loadable version of raw Purple Air synoptic data for practicing and code examples This dataset contains data for sensors in Washington and Oregon and was generated on 2023-08-12 by running:

library(AirSensor2)

initializeMazamaSpatialUtils()

source("global_vars.R") # contains PurpleAir_API_READ_KEY

example_pas_raw <-
  pas_downloadParseRawData(
    PurpleAir_API_READ_KEY,
    fields = PurpleAir_DATA_AVG_PM25_FIELDS,
    location_type = 0,
    modified_since = NULL,
    max_age = 3600 * 24,
    west = -125,
    east = -117,
    south = 42,
    north = 49
  )

save(example_pas_raw, file = "data/example_pas_raw.rda")

This dataset can be converted into a standard pas object with:

pas <- pas_enhanceRawData(example_pas_raw)

example_pas_raw

Format

A tibble with 2260 rows and 37 columns of data.

Source

https://www2.purpleair.com

See also