Initialize MazamaSpatialUtils package

Convenience function that wraps:

  data("SimpleCountriesEEZ", package = "MazamaSpatialUtils")
  data("SimpleTimezones", package = "MazamaSpatialUtils")

  MazamaSpatialUtils::setSpatialDataDir("~/Data/Spatial")
  MazamaSpatialUtils::loadSpatialData("NaturalEarthAdm1.rda")
  MazamaSpatialUtils::loadSpatialData("USCensusCounties.rda")

This function should be run before using pas_load(), as pas_load() uses the spatial data loaded by initializeMazamaSpatialUtils() to enhance raw synoptic data via pas_enhanceData().

If file logging is desired, these commands should be run individually with output log files specified as arguments to logger.setup() from the MazamaCoreUtils package.

initializeMazamaSpatialUtils(
  spatialDataDir = "~/Data/Spatial",
  stateCodeDataset = "NaturalEarthAdm1.rda",
  USCountiesDataset = "USCensusCounties.rda",
  logLevel = WARN
)

Arguments

spatialDataDir
Directory where MazamaSpatialUtils datasets are found.

stateCodeDataset
MazamaSpatialUtils dataset returning ISO 3166-2 . alpha-2 stateCodes.

USCountiesDataset
MazamaSpatialUtils dataset returning US county names.

logLevel
Logging level used if logging has not already been initialized.