Create a new PurpleAir synoptic dataset

Download, parse and enhance synoptic data from PurpleAir and return the results as a useful tibble with class purple_air_synoptic.

Steps include:

1) Download and parse synoptic data

2) Replace variable with more consistent, more human readable names.

3) Add spatial metadata for each sensor including:

timezone -- olson timezone

countryCode -- ISO 3166-1 alpha-2

stateCode -- ISO 3166-2 alpha-2

4) Convert data types from character to POSIXct and numeric.

Data requests are made with a bounding box determined from the required countryCodes parameter. If a single country is specified and additional stateCodes are specified, the bounding box will be limited to those states. Withing a single state, counties may be used to further limit the data.

pas_createNew(
  apiReadKey = NULL,
  countryCodes = NULL,
  stateCodes = NULL,
  counties = NULL,
  lookbackDays = 1,
  outsideOnly = TRUE,
  baseUrl = "https://api.purpleair.com/v1/sensors"
)

Arguments

apiReadKey
PurpleAir API Read Key.

countryCodes
ISO 3166-1 alpha-2 country codes used to subset the data. At least one countryCode must be specified.

stateCodes
ISO-3166-2 alpha-2 state codes used to subset the data. Specifying stateCodes is optional.

counties
US county names or 5-digit FIPS codes used to subset the data. Specifying counties is optional.

lookbackDays
Number of days to "look back" for valid data. Data are filtered to only include sensors with data more recent than lookbackDays ago.

outsideOnly
Logical specifying whether to restrict requests to outside sensors only.

baseUrl
Base URL for synoptic data.

Value

A PurpleAir Synoptic pas object.

See also

Examples

# \donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

initializeMazamaSpatialUtils()

pas <-
  pas_createNew(
    apiReadKey = API_READ_KEY,
    countryCodes = "US",
    stateCodes = "WA",
    counties = c("Okanogan", "Ferry"),
    lookbackDays = 1,
    outsideOnly = TRUE
  )

pas %>% pas_leaflet()

}, silent = FALSE)

#> Error in MazamaCoreUtils::stopIfNull(apiReadKey) : 
#>   object 'API_READ_KEY' not found

# }