Download synoptic data from PurpleAir with newer API

Download and parse synoptic data for the PurpleAir network of particulate sensors.

The synoptic data provides a view of the entire PurpleAir network and includes both metadata and recent PM2.5 averages for each deployed sensor.

pas_downloadParseRawData(
  apiReadKey = NULL,
  maxAge = 604800,
  outsideOnly = TRUE,
  west = NULL,
  east = NULL,
  south = NULL,
  north = NULL,
  baseUrl = "https://api.purpleair.com/v1/sensors"
)

Arguments

apiReadKey
PurpleAir API Read Key.

maxAge
Filter results to only include sensors modified or updated within the last number of seconds. Using a value of 0 will match sensors of any age.

outsideOnly
Logical specifying whether to restrict requests to outside sensors only.

west
Longitude of the western edge of the bounding box in which to find sensors.

east
Longitude of the eastern edge of the bounding box in which to find sensors.

south
Latitude of the southern edge of the bounding box in which to find sensors.

north
Latitude of the northern edge of the bounding box in which to find sensors.

baseUrl
Base URL for the PurpleAir API.

Value

Dataframe of synoptic PurpleAir data.

References

PurpleAir API

Examples

# \donttest{
# Fail gracefully if any resources are not available
try({

library(AirSensor2)

pas_raw <-
  pas_downloadParseRawData(
    API_READ_KEY,
    maxAge = 3600 * 24,
    outsideOnly = TRUE,
    west = -124,
    east = -123,
    south = 48,
    north = 50
  )

View(pas_raw[1:100,])

}, silent = FALSE)

#> Error in MazamaCoreUtils::stopIfNull(apiReadKey) : 
#>   object 'API_READ_KEY' not found

# }