Date filtering for PurpleAir Synoptic objects

Subsets a pas object by date.

Dates can be anything that is understood by MazamaCoreUtils::parseDatetime() including either of the following recommended formats:

"YYYYmmdd"

"YYYY-mm-dd"

Timezone determination precedence assumes that if you are passing in POSIXct values then you know what you are doing.

get timezone from startdate if it is POSIXct

use passed in timezone

pas_filterDate(pas = NULL, startdate = NULL, enddate = NULL, timezone = NULL)

Arguments

pas

PurpleAir Synoptic pas object.

startdate

Desired start datetime (ISO 8601).

enddate

Desired end datetime (ISO 8601).

timezone

Optional Olson timezone used to interpret dates.

Note

The incoming pas object must contain the date_created and last_modified fields. The returned pas object will retain sites where startdate <= last_seen && enddate >= pas$date_created.

See also

Examples

library(AirSensor2)

august_2018 <-
 example_pas %>%
  pas_filterDate(
    startdate = 20180701,
    enddate = 20181101,
    timezone = "America/Los_Angeles"
 )

fields <- c("sensor_index", "name", "date_created", "last_seen")
head(august_2018[,fields])
#> # A tibble: 6 × 4
#>   sensor_index name                      date_created        last_seen          
#>   <chr>        <chr>                     <dttm>              <dttm>             
#> 1 453          LRAPA-Oakridge City Hall  2016-11-07 18:24:07 2023-11-29 19:50:24
#> 2 2045         PSU STAR Lab Rose City P… 2017-07-12 18:08:42 2023-11-29 19:49:42
#> 3 2069         Ballard                   2017-07-12 18:52:26 2023-11-29 19:50:20
#> 4 2402         Enatai                    2017-07-31 20:27:30 2023-11-29 19:49:48
#> 5 2572         Rosewold                  2017-08-05 18:46:22 2023-11-29 19:51:07
#> 6 2651         Waterville                2017-08-08 18:14:55 2023-11-29 19:50:44